fix: Range-Bearing 수정
