+++ src/components/map/initialize/LoadSites.tsx
@@ -3,7 +3,7 @@ import { LatLngLiteral } from 'leaflet';
 import CustomAxios from 'module/axios';
 import { convertToWGS } from 'module/DMS';
 import React, { useEffect, useRef, useState } from 'react'
-import { LayerGroup, LayersControl, Marker, Pane, Rectangle, Tooltip } from 'react-leaflet';
+import { LayerGroup, LayersControl, Marker, Pane, Popup, Rectangle, Tooltip } from 'react-leaflet';
 
 
 const LoadEntireSite = async () => {
@@ -18,22 +18,13 @@ function LoadSites() {
         result.then(t => { setSite(t.data) }).catch(e => console.error(e))
     }, [])
 
-
-    useEffect(() => {
-        console.log(site)
-        if (site?.length) {
-            site.map(t => {
-                const { lat, lng } = t.siteCoordinate as LatLngLiteral
-            })
-        }
-    }, [site])
     return (
         <>
             {site?.map(t => {
                 const { lat, lng } = t.siteCoordinate as LatLngLiteral
-                console.log(t.siteName, lat,lng, '|', convertToWGS(lat), convertToWGS(lng) )
                 return <Marker key={t.siteId} position={{lat:convertToWGS(lat), lng:convertToWGS(lng)}} pane='site'>
                     <Tooltip>{t.siteName}</Tooltip>
+                    <Popup closeButton={false}>{t.siteName}</Popup>
                 </Marker>
             })}
         </>
